ARROW-10168: [Rust] [Parquet] Use Arrow schema by column too
Previously, if an Arrow schema was present in the Parquet metadata, that
schema would always be returned when requesting all columns via
`parquet_to_arrow_schema` and would never be returned when requesting a
subset of columns via `parquet_to_arrow_schema_by_columns`.

Now, if a valid Arrow schema is present in the Parquet metadata and a
subset of columns is requested by Parquet column index, the
`parquet_to_arrow_schema_by_columns` function will try to find a column
of the same name in the Arrow schema first, and then fall back to the
Parquet schema for that column if there isn't an Arrow Field for that
column.

This is part of what is needed to be able to restore Arrow types like
LargeUtf8 from Parquet.